Reintegrated Tooltip in core and simplified composition.
