+++ test/karma-files.json
@@ -18,7 +18,6 @@
 	"code/modules/exporting.src.js",
 	"code/modules/export-data.src.js",
 	"code/modules/parallel-coordinates.src.js",
-	"code/modules/tooltip.src.js",
 	"code/modules/accessibility.src.js",
 	"code/modules/annotations-advanced.src.js",
 	"code/modules/boost-canvas.src.js",
