+++ test/typescript-lint/tsconfig.json
@@ -213,9 +213,6 @@
         "../../code/modules/timeline.d.ts",
         "../../code/modules/timeline.src.d.ts",
 
-        "../../code/modules/tooltip.d.ts",
-        "../../code/modules/tooltip.src.d.ts",
-
         "../../code/modules/treegraph.d.ts",
         "../../code/modules/treegraph.src.d.ts",
